Extracted handling functionality for reuse
