+++ packages/browser-wallet-message-hub/src/index.ts
@@ -1,5 +1,6 @@
 export { Message, isMessage, isResponse, MessageResponse } from './messaging/message';
+export * from './messaging/handling';
+export * from './messaging/types';
 export { WalletMessageHandler } from './message-handlers/wallet-messagehandler';
 export { AbstractMessageHandler } from './message-handlers/abstract-messagehandler';
 export { InjectedMessageHandler } from './message-handlers/injected-messagehandler';
-export * from './message-handlers/types';
