+++ packages/browser-wallet-message-hub/src/messaging/types.ts
@@ -1,6 +1,3 @@
-// eslint-disable-next-line import/no-cycle
-import { Message } from './message';
-
 /**
  * Enumeration of the different types of messages that can be sent from the walletApi to the message handlers and vice versa
  */
@@ -28,9 +25,3 @@ export enum HandlerType {
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type Payload = any;
-
-export type EventHandler = (
-    message: Message,
-    respond: (payload: Payload) => void,
-    metadata?: chrome.runtime.MessageSender
-) => void;
