+++ packages/browser-wallet/src/inject/index.ts
@@ -6,13 +6,3 @@ import { IWalletApi, walletApi } from '@concordium/browser-wallet-api';
 (window as any).concordium = walletApi as IWalletApi;
 
 console.log('INJECT');
-
-// setInterval(
-//     () =>
-//         // eslint-disable-next-line @typescript-eslint/no-explicit-any
-//         (window as any).concordium.sendTransaction().then(() => {
-//             // eslint-disable-next-line no-console
-//             console.log('Promise resolved in Interval');
-//         }),
-//     10000
-// );
